qc report cellranger sub workflows tested

+++ utils/cloud.py
@@ -9,6 +9,7 @@ import tarfile
 import shutil
 import gzip
 
+
 class ResultsInterface(object):
 
     def __init__(self, version="", dir="./tmp"):
@@ -32,6 +33,7 @@ class ResultsInterface(object):
 class TenxDataStorage(object):
 
     def __init__(self, sampleid, version="v3"):
+        self.version = version
         self.sampleid = sampleid.replace(".tar.gz","")
         self.storage_account = "scrnadata"
         self.container = "cellranger{}".format(version)
@@ -56,18 +58,16 @@ class TenxDataStorage(object):
         bam_tarball_name = os.path.split(bam_tarball)[-1]
         outs_tarball = tenx.outs_tarball()
         outs_tarball_name = os.path.split(outs_tarball)[-1]
-        bam_blob_path = "{sampleid}/{local}".format(self.sampleid, bam_tarball_name)
-        self.upload("bams" ,bam_blob_path, bam_tarball)
-        outs_blob_path = "{sampleid}/{local}".format(self.sampleid, bam_tarball_name)
-        self.upload("cellranger", outs_blob_path, bam_tarball)
+        self.upload("bams" ,bam_tarball_name, bam_tarball)
+        self.upload("cellranger{}".format(self.version), outs_tarball_name, outs_tarball)
 
 
     def unpack(self,path):
         tar = tarfile.open(path)
-        tar.extractall(".cache")
+        tar.extractall(".cache/{}".format(self.sampleid))
         tar.close()
-        assert os.path.exists(".cache/outs"), "Failed to download."
-        os.rename(".cache/outs",os.path.join(self.cache,self.sampleid))
+        if os.path.exists(".cache/outs"):
+            os.rename(".cache/outs",os.path.join(self.cache,self.sampleid))
 
 
     def download(self):
@@ -80,9 +80,24 @@ class TenxDataStorage(object):
         return self.tenx_path
 
     def upload(self,container, blob, local):
-        print ("Uploading {} to {} in {}".format(container,local,blob))
-        cmd = "az storage blob upload --container-name {} --file {} --name {} --auth-mode login".format(container,local,blob)
-        subprocess.call(cmd.split())
+        print ("Uploading {} to {} in {}".format(container,local,container))
+        self.block_blob_service.create_blob_from_path(container, blob, local)
+
+class ReportStorage(object):
+
+    def __init__(self, results):
+        self.results = results
+        self.storage_account = "scrnadata"
+        self.container = "reports"
+        self.block_blob_service = BlockBlobService(account_name=self.storage_account, sas_token='?sv=2018-03-28&ss=bfqt&srt=sco&sp=rwdlacup&se=2021-03-19T02:52:48Z&st=2019-02-22T19:52:48Z&spr=https&sig=4oAGvIyqi9LPY89t21iWWp4XbbIgpmaldgcwWUOuf14%3D')
+
+    def upload(self, path, sampleid):
+        reportdir = self.results.report_dir
+        reporttarball = os.path.join(path, "{}.tar.gz".format(sampleid))
+        with tarfile.open(reporttarball, "w:gz") as tar:
+            tar.add(reportdir, arcname=os.path.basename(reporttarball))
+        blobtarball = os.path.split(reporttarball)[1]
+        self.block_blob_service.create_blob_from_path(self.container, blobtarball, reporttarball)
 
 class FastqDataStorage(object):
 
@@ -127,8 +142,8 @@ class ReferenceDataStorage(object):
         tar.close()
 
     def download(self):
-        if not os.path.exists(self.reference):
-            local = os.path.join(self.referencepath, "{}.tar.gz".format(self.build))
+        local = os.path.join(self.referencepath, "{}.tar.gz".format(self.build))
+        if not os.path.exists(local):
             self.block_blob_service.get_blob_to_path(self.container, "{}.tar.gz".format(self.build), local)
             self.extract(local)
         return self.reference
